requirements.txt
# Binance Futures Trading Bot Requirements

# Core dependencies
python-binance==1.0.19
requests==2.31.0

# Environment management
python-dotenv==1.0.0

# Optional but recommended
colorama==0.4.6  # For colored terminal output on Windows
